Have been playing around with the parser trying to get it to work over multiple threads.
Sadly the library won't work well with multiple threads when parsing the ticks resulting in no time gain.

This would be solvable if we added the library from the git repo and set a lock on the part of the code that works with the same code. In that case we could potentially get the parser to work 10 times as fast.